corrections et modificatiton facture achat et vente

+++ Dtos/LigneFactureDto.cs
@@ -9,7 +9,7 @@ namespace Administration.Dtos
         public int ID_Service { get; set; }
         [AllowNull]
         public int? ID_FactureVente { get; set; }
-        public int Quantite { get; set; }
+        public decimal Quantite { get; set; }
         public decimal Total_LigneFV { get; set; }
         public decimal Total_LigneHT { get; set; }
     }
